% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-analysis.R
\name{aop_network}
\alias{aop_network}
\title{Create Adverse Outcome Pathway (AOP) Network data}
\usage{
aop_network(aop_df, dise_voc, weight_column)
}
\arguments{
\item{aop_df}{A data frame containing AOP data with columns for the right-hand side (rhs)
and left-hand side (lhs) of the associations.}

\item{dise_voc}{A data frame containing the disease vocabulary with columns \code{DiseaseID}, \code{DiseaseName},
and \code{DiseaseGroup}.}

\item{weight_column}{A string specifying the column in \code{aop_df} to use as the weight for the edges.}
}
\value{
A list with two components:
\item{edges}{A data frame representing the edges of the network, with columns \code{from}, \code{to}, and \code{weight}.}
\item{vertices}{A data frame representing the vertices of the network, with columns \code{name}, \code{nodes}, and \code{size}.}
}
\description{
The \code{aop_network()} function constructs a network of Adverse Outcome Pathways (AOPs)
from the provided AOP data frame and disease vocabulary. It generates a network graph
with nodes representing diseases and genes, and edges representing the connections
between them, weighted by a specified column.
}
\examples{
\dontrun{
# Sample data frames
aop_df <- data.frame(lhs = c("Gene1", "Gene2"), rhs = c("Disease1", "Disease2"), support = c(0.8, 0.6))
dise_voc <- data.frame(DiseaseID = c("Disease1", "Disease2"), DiseaseName = c("Disease A", "Disease B"), DiseaseGroup = c("Group1", "Group2"))

# Create AOP network
aop_network_data <- aop_network(aop_df, dise_voc, "support")
edges <- aop_network_data$edges
vertices <- aop_network_data$vertices
}
}
