% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{cel2exprs}
\alias{cel2exprs}
\title{Convert Affymetrix CEL Files to Gene Expression Matrix}
\usage{
cel2exprs(cel_files, organism = "rat", method = "rma", ...)
}
\arguments{
\item{cel_files}{A vector containing the full paths of CEL files.}

\item{organism}{The sample organism used in the experiment, either \code{"human"} or \code{"rat"}.}

\item{method}{The normalization method to use, either \code{"rma"} or \code{"mas5"}. Default is \code{"rma"}.}

\item{...}{Additional options for processing CEL files, passed on to
\link[affy:rma]{rma} or \code{\link[affy:mas5]{mas5}}, such as \code{normalize}, \code{background}, etc.}
}
\value{
A matrix of raw gene expression data:
\itemize{
\item Rows represent probes
\item Columns represent sample CEL IDs
}
}
\description{
The \code{cel2exprs()} function extracts gene expression data from CEL files, outputting
an expression matrix. The dimensions of the matrix depend on the number of CEL files
and the probes specified in the Chip Definition File (CDF).

This function utilizes core functions \code{\link[affy:mas5]{mas5}} and \code{\link[affy:rma]{rma}}
from the affy package to quantify mRNA intensity of probes/genes. The expression
measurements are provided on a log base 2 scale.
}
\examples{
\dontrun{
file_path <- tggates_rawdata(compound = "2-nitrofluorene")
cel_file <- file.path(file_path, "celfiles")
cel_files <- list.files(cel_file, full.names = TRUE)
expr_data <- cel2exprs(cel_files)
head(expr_data)
}
}
\seealso{
\code{\link[affy:mas5]{affy::mas5()}} for Affymetrix version 5 (MAS5) model,
\code{\link[affy:rma]{affy::rma()}} for Robust Multi-array Average (RMA) model
}
\concept{Gene Expression Data Processing}
