% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_diff}
\alias{col_diff}
\title{Compute pairwise difference between matrix columns}
\usage{
col_diff(avg_data)
}
\arguments{
\item{x}{A data matrix of size n times p. Where rows are observations and
columns are features.}
}
\value{
A matrix of size n times (p choose 2), where each column is the
difference between two of the original columns.
}
\description{
Compute pairwise difference between matrix columns
}
\examples{
n = 10
p = 3
x = matrix(rnorm(n*p), nrow = n, ncol = p, byrow = TRUE)
rownames(x) <- paste0("r", 1:n)
colnames(x) <- paste0("c", 1:p)
col_diff(x)
}
