% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{ctd_link}
\alias{ctd_link}
\title{Download Link for Data in the CTD Database}
\usage{
ctd_link(
  compounds,
  association = c("cgixns", "genes_curated", "genes_inferred", "diseases",
    "diseases_curated", "diseases_inferred"),
  action_type = "ANY"
)
}
\arguments{
\item{compounds}{A character vector specifying the names of the compounds.}

\item{association}{A character string specifying the type of association to retrieve (e.g., "genes_curated"). Default is \code{"cgixns"}.}

\item{action_type}{A character string specifying the type of action (e.g., "ANY"). Default is \code{"ANY"}.}
}
\value{
A character vector containing the download links for the specified compounds.
}
\description{
The \code{ctd_link()} function retrieves the download link for perturbation data of compounds
from the \href{http://ctdbase.org/}{CTD (Comparative Toxicogenomics Database)}.
}
\examples{
# Downloading link for association data of Ethanol and curated genes
ctd_link(compounds = "Ethanol", association = "genes_curated")
# Downloading link for association data of all type between Ethanol and Esrriol, and genes
ctd_link(compounds = c("Ethanol", "Estriol"), association = "cgixns")
}
\concept{CTD data download helpers}
