% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure&summary-matrix.R
\name{data_str}
\alias{data_str}
\title{Structure of experimental data for analyzing targeted toxicity}
\usage{
data_str(..., metadata, error_call = caller_env())
}
\arguments{
\item{...}{Either multiple vectors of individual compound names or multiple vectors of
compound names grouped in a list.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}
}
\value{
A list total levels in different stage or a class of \code{ToxAssay} object .
\itemize{
\item group
\item compound
\item dose
\item time
\item replication
}
}
\description{
The \code{data_str()} function finds the number of levels in dose, time point, and
replication for a list of compounds in different groups. \code{data_str()} uses the
function \code{data_str()} to determine the structure of levels in compounds and
concatenates them to create the data structure for multiple compounds in a group.
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
data_str(gr, metadata = sim_data$metadata)
}
\seealso{
\code{\link[=compound_str]{compound_str()}} for getting structure of a compound
}
