% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm-degs.R
\name{de_genes}
\alias{de_genes}
\title{Identification of initial differentially expressed genes}
\usage{
de_genes(
  ...,
  ge_matrix,
  metadata,
  p_cutoff = 0.05,
  p_adjust = "none",
  gr_diff = FALSE,
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{p_cutoff}{P-value cutoff threshold.}

\item{p_adjust}{Adjusted \code{p-values} using Bonferroni methods. default is \code{TRUE}.}

\item{gr_diff}{Show the column(s) for group mean differences. The default is \code{FALSE}.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
An object of \code{data.frame} or \code{tibble} about genes table.
}
\description{
The \code{de_genes()} function identify initial differentially expressed genes using the HLM.
}
\details{
\code{de_genes()} identifies genes that exhibit statistically significant differential
expression across two or more query compound groups. A gene is deemed statistically
significant if the \emph{p-value} is significant in HLM.
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
de_genes(gr, ge_matrix = sim_data$expression, metadata = sim_data$metadata)
}
