% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{direct_sum}
\alias{direct_sum}
\title{Direct sum of two vectors or matrices}
\usage{
direct_sum(A, B)
}
\arguments{
\item{A}{A numeric matrix or vector}

\item{B}{A numeric matrix or vector}
}
\value{
a matrix of direct sum
}
\description{
The function ``
}
\examples{
A <- matrix(1, nrow = 3, ncol = 3)
B <- matrix(2, nrow = 3, ncol = 3)
direct_sum(A, B)
}
