% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-analysis.R
\name{disease_similarity}
\alias{disease_similarity}
\title{Calculate Disease Similarity Based on Gene Overlap}
\usage{
disease_similarity(
  disease_data,
  gene_column,
  target_column,
  genes = NULL,
  condition = TRUE
)
}
\arguments{
\item{disease_data}{A data frame containing disease information.}

\item{gene_column}{The column name in \code{disease_data} that contains gene symbols.}

\item{target_column}{The column name in \code{disease_data} that contains the target or disease information.}

\item{genes}{An optional vector of genes to filter the data. If NULL, all genes in the data are used.}

\item{condition}{An optional condition to filter the data. Default is TRUE.}
}
\value{
A similarity matrix where each entry (i, j) represents the similarity between disease i and disease j.
}
\description{
This function calculates the similarity between diseases based on the overlap of associated genes.
}
\examples{
chemical_name <- paste0("Compound", 1:10)
gene_symbol <- paste0("GENE", 1:30)
disease_id <- paste0("Disease", 1:30)
compound_gene <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                            gene_symbol = sample(gene_symbol, 100, replace = TRUE),
                            organism_id = sample(c(9606, 10116), 100, replace = TRUE)) \%>\%
                            dplyr::distinct()
compound_disease <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                               disease_id = sample(disease_id, 100, replace = TRUE)) \%>\%
                               dplyr::distinct()
trans_data <- get_transaction(compound_gene, compound_disease)
aop_data <- get_aops(transaction = trans_data[[1]], genes = trans_data[[2]], diseases = trans_data[[3]])
disease_similarity(aop_data, gene_column = lhs, target_column = rhs)
}
