% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{drugmatrix_dictionary}
\alias{drugmatrix_dictionary}
\title{Collect Metadata for Compounds in the DrugMatrix Database}
\usage{
drugmatrix_dictionary(
  compounds,
  tissue = c("Liver", "Kidney", "Heart", "Hepatocytes")
)
}
\arguments{
\item{compounds}{A vector of compound names to query.}

\item{tissue}{The tissue type used to generate perturbation data. Default is \code{Liver}.
Refer to the \strong{Available data for different tissue} section for more details.}
}
\value{
A data frame containing the metadata.
}
\description{
The \code{dm_metadata()} function retrieves metadata for query compounds from the
\href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} database. It uses the compound name and tissue type to
obtain Affymetrix CEL images and their attribute data from the
\href{https://www.ncbi.nlm.nih.gov/geo/}{GEO} database. Each CEL file represents raw expression
data of \code{probes} measured at various dose and time points, or combinations thereof, from chemical compound exposure.
}
\section{Available Data for Different Tissues}{

The \href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} database includes tests on
over 600 distinct compounds, conducted in both \verb{in vivo} and \verb{in vitro}
experimental setups. Male Sprague Dawley rats were used in both \verb{in vivo} and \verb{in vitro}
experiments, with \verb{in vitro} studies specifically employing rat primary hepatocytes.
The available data for different tissues are:
\itemize{
\item Liver
\item Kidney
\item Heart
\item Hepatocytes (Rat Primary Hepatocytes)
}
}

\examples{
# Perturbation data of Ethanol from liver tissue
drugmatrix_dictionary("Ethanol", "Liver")
# Perturbation data of Ethanol from Kidney tissue
drugmatrix_dictionary("Ethanol", "Kidney")
}
