% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{drugmatrix_link}
\alias{drugmatrix_link}
\title{Download Links for Data in the DrugMatrix Database}
\usage{
drugmatrix_link(
  compounds,
  tissue = c("Liver", "Kidney", "Heart", "Hepatocytes")
)
}
\arguments{
\item{compounds}{A vector of compound names to query.}

\item{tissue}{The tissue type used to generate perturbation data. Default is \code{Liver}.
Refer to the \strong{Available data for different tissue} section for more details.}
}
\value{
A vector of download links for the specified compounds and tissue.
}
\description{
The \code{drugmatrix_link()} function retrieves download links for the perturbation data
of specified compounds from the \href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} database.
}
\section{Available Data for Different Tissues}{

The \href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} database includes tests on
over 600 distinct compounds, conducted in both \verb{in vivo} and \verb{in vitro}
experimental setups. Male Sprague Dawley rats were used in both \verb{in vivo} and \verb{in vitro}
experiments, with \verb{in vitro} studies specifically employing rat primary hepatocytes.
The available data for different tissues are:
\itemize{
\item Liver
\item Kidney
\item Heart
\item Hepatocytes (Rat Primary Hepatocytes)
}
}

\examples{
# Downloading link for perturbation data of Ethanol from Liver tissue
drugmatrix_link(compounds= "Ethanol", tissue = "Liver")
# Downloading link for perturbation data of Ethanol and Esrriol from Liver tissue
drugmatrix_link(c("Ethanol", "Estriol"))
}
\seealso{
Other DrugMatrix data download helpers: 
\code{\link{drugmatrix_compounds}()},
\code{\link{drugmatrix_rawdata}()}
}
\concept{DrugMatrix data download helpers}
