% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{drugmatrix_rawdata}
\alias{drugmatrix_rawdata}
\title{Download Microarray and Metadata for Query Compounds from the DrugMatrix Database}
\usage{
drugmatrix_rawdata(
  compounds,
  tissue = c("liver", "kidney", "heart", "hepatocytes"),
  output_dir = missing_arg()
)
}
\arguments{
\item{compounds}{A vector of compound names to query.}

\item{tissue}{The tissue type used to generate perturbation data. Default is \code{Liver}.
Refer to the \strong{Available data for different tissue} section for more details.}

\item{output_dir}{Output directory for downloading data. If not specified,
the temporary path of the current R session will be used.}
}
\value{
A folder containing \code{CEL} files and a \code{TSV} metadata file.
}
\description{
The \code{drugmatrix_rawdata()} function downloads raw Affymetrix CEL files and their associated
attribute files (data dictionary) for a specified compound from the
\href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} database. A folder named
\strong{celfiles} will be created in your specified output directory to store the CEL image
files, and an accompanying TSV file containing metadata will also be generated in the
parent folder.

This function retrieves Affymetrix CEL image files for specific compounds from the
DrugMatrix database, storing all downloaded data in a user-selected directory. Since gene
expression is measured at various doses and time points or combinations thereof, a
collection of CEL files will be downloaded into a \strong{celfiles} folder. Additionally, a
TSV file containing comprehensive metadata will be saved in the parent directory. Note
that download speed depends on your internet connection.
}
\examples{
compounds <- "Ethanol"
\dontrun{
drugmatrix_rawdata(compounds, "Liver")
}
}
\seealso{
Other DrugMatrix data download helpers: 
\code{\link{drugmatrix_compounds}()},
\code{\link{drugmatrix_link}()}
}
\concept{DrugMatrix data download helpers}
