% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{expand_data}
\alias{expand_data}
\title{Expand existing data by adding data for new compounds}
\usage{
expand_data(
  compounds,
  ge_matrix,
  metadata,
  multicore = FALSE,
  store = FALSE,
  method = "rma",
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{compounds}{A character vector of compound names to retrieve from the TG-GATES database.}

\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{method}{A character string specifying the preprocessing method to use. Default is \code{"rma"}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}

\item{database}{The database to be used, either \code{tggates} or \code{drugmatrix}.}
}
\value{
A list.
\itemize{
\item A matrix of expanded gene expression data of compounds, probes in row and sample
(CEL-ID) in column.
\item A data frame of expanded metadata about sample and phenotypic information of compounds.
}
}
\description{
The \code{expand_data()} function adds data for extra compounds to the existing data.

The function \code{expand_data()} is used to add data for one or more compounds to the
existing data. It processes new compound data and updates both \code{expression} and
\code{metadata} based on the database used. You can add compounds from both the
\href{https://toxico.nibiohn.go.jp/english/}{TG-GATEs} and
\href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} databases. The available
compounds in TG-GATEs and DrugMatrix can be found by calling the functions
\code{\link[=tggates_compounds]{tggates_compounds()}} and \code{\link[=dm_compounds]{dm_compounds()}}, respectively.
}
\examples{
\dontrun{
otg_data <- get_tggates(compounds = c("2-nitrofluorene", "N-nitrosomorpholine"))
otg_expndata <- expand_data(compounds = c("2-nitrofluorene", "N-nitrosomorpholine", "N-methyl-N-nitrosourea"),
                            ge_matrix = otg_data$expression,
                            metadata = otg_data$metadata)
head(otg_expndata$expression)
head(otg_expndata$metadata)
}
}
\concept{Data update helpers}
