% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-analysis.R
\name{gene_similarity}
\alias{gene_similarity}
\title{Calculate gene Similarity Based on Gene Overlap}
\usage{
gene_similarity(
  gene_data,
  disease_column,
  target_column,
  diseases = NULL,
  condition = TRUE
)
}
\arguments{
\item{target_column}{The column name in \code{disease_data} that contains the target or disease information.}

\item{condition}{An optional condition to filter the data. Default is TRUE.}

\item{link_data}{A data frame containing gene and disease information.}

\item{gene_column}{The column name in \code{disease_data} that contains gene symbols.}

\item{genes}{An optional vector of genes to filter the data. If NULL, all genes in the data are used.}
}
\value{
A similarity matrix where each entry (i, j) represents the similarity between disease i and disease j.
}
\description{
This function calculates the similarity between genes based on the overlap of associated diseases.
}
\examples{
chemical_name <- paste0("Compound", 1:10)
gene_symbol <- paste0("GENE", 1:30)
disease_id <- paste0("Disease", 1:30)
compound_gene <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                            gene_symbol = sample(gene_symbol, 100, replace = TRUE),
                            organism_id = sample(c(9606, 10116), 100, replace = TRUE)) \%>\%
                            dplyr::distinct()
compound_disease <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                               disease_id = sample(disease_id, 100, replace = TRUE)) \%>\%
                               dplyr::distinct()
trans_data <- get_transaction(compound_gene, compound_disease)
}
