% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset&average-data.R
\name{get_avgFC}
\alias{get_avgFC}
\title{Average gene expression in different levels}
\usage{
get_avgFC(
  ...,
  ge_matrix,
  metadata,
  probe_id,
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{probe_id}{The gene/probe ID.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list of average data
}
\description{
The function \code{get_avgFC()} is used to get average gene expression value in \code{group},
\code{compound}, \code{dose} and \code{time} stages.
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
avg_value <- get_avgFC(gr, ge_matrix = sim_data$expression, metadata = sim_data$metadata, probe_id = "DE1")
}
