% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{get_ctd}
\alias{get_ctd}
\title{Retrieve CTD Data for Compounds}
\usage{
get_ctd(
  compounds,
  gene_parm = "cgixns",
  disease_parm = "diseases",
  action_type = "ANY",
  store = FALSE,
  output_dir = missing_arg()
)
}
\arguments{
\item{compounds}{A character vector specifying the names of the compounds.}

\item{gene_parm}{A character string specifying the gene association parameter. Default is \code{"cgixns"}.}

\item{disease_parm}{A character string specifying the disease association parameter. Default is \code{"diseases"}.}

\item{action_type}{A character string specifying the type of action (e.g., "ANY"). Default is \code{"ANY"}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list with the following components:
\itemize{
\item \code{compound_gene}: A data frame containing the gene associations for the specified compounds.
\item \code{compound_disease}: A data frame containing the disease associations for the specified compounds.
}
}
\description{
The \code{get_ctd()} function retrieves data for specified compounds from the
\href{http://ctdbase.org/}{Comparative Toxicogenomics Database (CTD)}, including gene and disease associations.
}
\examples{
\dontrun{
ctd_data <- get_ctd(compounds = c("Ethanol", "Estriol"), gene_parm = "cgixns", disease_parm = "diseases")
head(ctd_data$compound_gene)
head(ctd_data$compound_disease)
}
}
\concept{CTD data retrieval helpers}
