% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{get_drugmatrix}
\alias{get_drugmatrix}
\title{Retrieve and Process DrugMatrix Data}
\usage{
get_drugmatrix(
  compounds,
  tissue = "Liver",
  fc = TRUE,
  method = "rma",
  store = FALSE,
  multicore = FALSE,
  output_dir = missing_arg(),
  ...
)
}
\arguments{
\item{compounds}{A character vector specifying the names of the compounds to retrieve.}

\item{tissue}{A character string specifying the tissue type. Default is \code{"Liver"}.}

\item{fc}{A logical value indicating whether to apply fold change. Default is \code{TRUE}.}

\item{method}{A character string specifying the method for processing the CEL files. Default is \code{"rma"}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}

\item{...}{Additional arguments passed to the \code{process_drugmatrix} function.}
}
\value{
A list with the following components:
\itemize{
\item \code{expression}: A matrix of gene expression values with probes in rows and samples (CEL IDs) in columns.
\item \code{metadata}: A data frame containing metadata about the samples.
}
}
\description{
The \code{get_dm()} function retrieves raw data files for specified compounds and tissue types
from the DrugMatrix database and processes them to produce gene expression data and metadata.
}
\examples{
\dontrun{
dm_data <- get_drugmatrix(compounds = "Ethanol", tissue = "Liver")
head(dm_data$expression)
head(dm_data$metadata)
}
}
