% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-enrichment.R
\name{get_enrichment}
\alias{get_enrichment}
\title{Get Gene Enrichment Results}
\usage{
get_enrichment(
  gene_df,
  category = c("KEGG", "Process", "Component", "Function", "RCTM", "WikiPathways"),
  path_n = 10,
  organism = c("rat", "human"),
  score_threshold = 200,
  version = "12"
)
}
\arguments{
\item{gene_df}{A data frame containing the gene information with a column named \code{gene_symbol}.}

\item{category}{A character vector specifying the categories for enrichment. Options are "KEGG" (the default), "Process", "Component", "Function", "RCTM", and "WikiPathways".}

\item{path_n}{An integer specifying the number of top pathways to return (default is 10).}

\item{organism}{A character string specifying the organism. Options are "rat" or "human".}

\item{score_threshold}{A numeric value specifying the score threshold for STRING database interactions (default is 200).}

\item{version}{A character string specifying the version of the STRING database (default is "12").}
}
\value{
A tibble containing enriched pathways with columns:
\item{ID}{The term ID of the pathway.}
\item{pathway}{The name of the pathway.}
\item{n_genes}{The number of genes in the pathway.}
\item{p_value}{The p-value of the enrichment.}
\item{fdr}{The false discovery rate of the enrichment.}
\item{genes}{The genes involved in the pathway.}
}
\description{
The \code{get_enrichment} function enriches a set of genes based on specified categories such as KEGG pathways,
Gene Ontology (GO) biological processes, molecular functions, cellular components, Reactome, and WikiPathways.
It maps the genes to the STRING database, retrieves enrichment results, and processes the results to return a data frame of enriched pathways.
}
\examples{
\dontrun{
# Sample gene data frame
gene_df <- data.frame(gene_symbol = c("Gene1", "Gene2", "Gene3"))

# Get enrichment results using KEGG pathways for humans
enriched_genes <- get_enrichment(gene_df, category = "KEGG", path_n = 5, organism = "human")

# View the enriched pathways
print(enriched_genes)
}
}
