% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structure&summary-matrix.R
\name{get_matrix}
\alias{get_matrix}
\title{Quadratic matrices for F-statistics}
\usage{
get_matrix(expr_str)
}
\arguments{
\item{expr_str}{The structure of the experiment, either a list detailing the total
number of labels at each level or a \code{ToxAssay} object representing the data
structure, can be obtained using the \code{\link[=data_str]{data_str()}} function.}
}
\value{
a list of design matrices and others information in the experiment
\itemize{
\item \code{Fstat_A}, quadratic matrix A
\item \code{BFstat_A}, quadratic matrix B
\item \code{group_mat}, summary matrix used to calculate average expression at group level
\item \code{compound_mat}, summary matrix used to calculate average expression at compound level
\item \code{dose_mat}, summary matrix used to calculate average expression at dose level
\item \code{time_mat}, summary matrix used to calculate average expression at time level
}
}
\description{
The \code{get_quadmat()} function calculates the quadratic matrices \strong{A} and \strong{B}
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
sim_str <- data_str(gr, metadata = sim_data$metadata)
summary_mat <- get_matrix(sim_str)
}
