% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi-network.R
\name{get_netdata}
\alias{get_netdata}
\title{Get network data}
\usage{
get_netdata(
  gene_df,
  organism = c("human", "rat"),
  cluster_method = c("edge.betweenness", "fastgreedy", "walktrap", "spinglass"),
  n_percent = 5,
  score_threshold = 200,
  version = "12"
)
}
\arguments{
\item{gene_df}{A data frame of gene information or a \code{tgxtool} class of object.}

\item{organism}{A character string specifying the sample organism used for the experiment. Options are "human" or "rat". Default is "rat".}

\item{cluster_method}{Cluster algorithm to find community. You can choose between "fastgreedy",
"walktrap", "spinglass" and "edge.betweenness" (the default is \code{NULL}).}

\item{score_threshold}{A numeric value specifying the threshold for the combined scores of the interactions. Default is 200.}

\item{version}{A character string specifying the version of the STRING database. Default is "11.5".}

\item{gene_col}{The name of the column for gene name in \code{gene_df}.}

\item{p.value_col}{The name of the column for p-values in \code{gene_df}.}
}
\value{
A list contain two data frames
\enumerate{
\item \code{vertices}: A data frame of information about nodes.
\item \code{edges}: A data frame of information about edges.
}
}
\description{
The function \code{get_netdata()} is used to extract network data from gene table.
}
\examples{
\dontrun{
net_data <- get_netdata(
  gene_data,
  gene_col = "gene_name",
  p.value_col = "group",
  organism = "rat",
  score_threshold = 400,
  cluster_method = "edge.betweenness"
)
}
}
