% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset&average-data.R
\name{get_subset}
\alias{get_subset}
\title{Get a subset of gene expression data based on selected parameters}
\usage{
get_subset(
  ...,
  ge_matrix,
  metadata,
  probes = NULL,
  dose = NULL,
  time = NULL,
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{probes}{A vector of Affymetrix probe IDs to be included. The default is \code{NULL},
which includes all probes..}

\item{dose}{A character vector of dose levels to subset the metadata and expression matrix. If NULL, all dose levels will be used.}

\item{time}{A character vector of time points to subset the metadata and expression matrix. If NULL, all time points will be used.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list with the following components:
\item{expression}{A matrix of gene expression data that has been subset based on the specified dose and time levels, and optionally the selected probes.}
\item{metadata}{A data frame containing metadata corresponding to the samples in the subsetted expression matrix, including dose, time, and compound information.}
}
\description{
This function subsets gene expression data (\code{ge_matrix}) and metadata (\code{metadata}) based on selected
dose levels, time points, and other parameters. It performs error handling for incorrect dose
or time inputs and offers options for parallel processing and data storage.
}
\details{
The function first checks for valid dose and time levels in the \code{metadata} and throws an error if
any mismatch is found. It then subsets the gene expression data (\code{ge_matrix}) and metadata
(\code{metadata}) based on the specified dose and time levels. Parallel processing is optionally supported via
the \code{multicore} parameter, and the results can be stored to a specified directory if \code{store} is TRUE.
}
\examples{
sim_data <- simulate_tgxdata()
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
sub_data <- get_subset(gr, ge_matrix = sim_data$expression, metadata = sim_data$metadata, probes = paste0("DE", 1:10), dose = "Dose1", time = "Time1")
}
