% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop-analysis.R
\name{get_transaction}
\alias{get_transaction}
\title{Get Transaction Data for Compounds, Genes, and Diseases}
\usage{
get_transaction(
  compound_gene,
  compound_disease,
  compounds = NULL,
  genes = NULL,
  diseases = NULL
)
}
\arguments{
\item{compound_gene}{A data frame containing chemical and gene information.
It should have columns \code{ChemicalName}, \code{GeneSymbol}, and \code{OrganismID}, similar to the CTD database.
One can download this data using the function \code{get_ctd()}.}

\item{compound_disease}{A data frame containing chemical and disease information.
It should have columns \code{ChemicalID} and \code{DiseaseID}, similar to the CTD database.
One can download this data using the function \code{get_ctd()}.}

\item{compounds}{An optional character vector of compound names to filter.
If \code{NULL} (default), all compounds from \code{compound_gene} are used.}

\item{gene_id}{An optional character vector of gene IDs to filter.
If \code{NULL} (default), all gene IDs from \code{compound_gene} are used.}

\item{disease_id}{An optional character vector of disease IDs to filter.
If \code{NULL} (default), all disease IDs from \code{compound_disease} are used.}
}
\value{
A list with three components:
\item{bin_data}{A binary matrix indicating the presence of interactions between chemicals, genes, and diseases.}
\item{genes}{A vector of gene symbols included in the binary matrix.}
\item{diseases}{A vector of disease IDs included in the binary matrix.}
}
\description{
The \code{get_transaction()} function processes and retrieves transaction data for specified compounds,
genes, and diseases. It takes two data frames as input and filters the data based on the provided
compounds, gene IDs, and disease IDs. The function merges the gene and disease data and returns
a binary matrix indicating the presence of interactions between chemicals, genes, and diseases.
}
\examples{
chemical_name <- paste0("Compound", 1:10)
gene_symbol <- paste0("GENE", 1:30)
disease_id <- paste0("Disease", 1:30)
compound_gene <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                            gene_symbol = sample(gene_symbol, 100, replace = TRUE),
                            organism_id = sample(c(9606, 10116), 100, replace = TRUE)) \%>\%
                            dplyr::distinct()
compound_disease <- data.frame(chemical_name = sample(chemical_name, 100, replace = TRUE),
                               disease_id = sample(disease_id, 100, replace = TRUE)) \%>\%
                               dplyr::distinct()
get_transaction(compound_gene, compound_disease)
}
