% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm-degs.R
\name{hlm}
\alias{hlm}
\title{Hierarchical linear model}
\usage{
hlm(Y, A, B, a, p, n, error_call = rlang::caller_env())
}
\arguments{
\item{Y}{Gene expression data matrix, probes in row and samples in column.}

\item{A}{quadratic matrix A.}

\item{B}{quadratic matrix B.}

\item{a}{Number of compound groups.}

\item{p}{Number of possible cluster in the data.}

\item{n}{Total number of samples in the data.}
}
\value{
A vector of p-values with a size equal to the number of rows in \strong{Y}.
}
\description{
The \code{hlm()} function models gene expression data utilizing the hierarchical linear
model (HLM).
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
sim_str <- data_str(gr, metadata = sim_data$metadata)
summary_mat <- get_matrix(sim_str)
hlm(Y = sim_data$expression, A = summary_mat$A, B =summary_mat$B, a = sim_str[[1]], p = sum(sim_str[[4]]), n = sum(sim_str[[5]]))
}
