% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-check.R
\name{is_compound}
\alias{is_compound}
\title{Validate Compound Names Against Databases}
\usage{
is_compound(
  compounds,
  database = rlang::missing_arg(),
  species = rlang::missing_arg(),
  data_type = rlang::missing_arg(),
  tissue = rlang::missing_arg(),
  dose_type = rlang::missing_arg(),
  error_call = rlang::caller_env()
)
}
\arguments{
\item{compounds}{A character vector of compound names to be checked.}

\item{database}{The name of the database to check against. Should be either \code{"tggates"} or \code{"drugmatrix"}. If not specified, both databases are checked.}

\item{species}{The species to filter for when using the \code{"tggates"} database. Should be \code{"Rat"} or \code{"Human"}.}

\item{data_type}{The type of data to filter for when using the \code{"tggates"} database. Should be \code{"in_vivo"} or \code{"in_vitro"}.}

\item{tissue}{The tissue type to filter for. For \code{"tggates"}, it should be \code{"Liver"} or \code{"Kidney"}. For \code{"drugmatrix"}, it should be \code{"Liver"}, \code{"Kidney"}, \code{"Heart"}, or \code{"Hepatocytes"}.}

\item{dose_type}{The dose type to filter for when using the \code{"tggates"} database. Should be \code{"Single"} or \code{"Repeat"}.}

\item{error_call}{Environment used to capture errors. Default is \code{caller_env()}.}
}
\value{
Returns \code{TRUE} if all compound names are valid, otherwise throws an error.
}
\description{
This function checks whether the provided compound names exist in the specified database
(TG-GATEs or DrugMatrix) and validates additional parameters such as species, data type,
tissue, and dose type. If no database is specified, it checks the compounds against both
the TG-GATEs and DrugMatrix databases.
}
\examples{
# Check if a compound is in the TG-GATEs or DrugMatrix database
is_compound(compounds = c("aspirin", "acetaminophen"), database = "tggates", species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "Single")
\dontrun{
# Check if a compound exists in either the TG-GATEs or DrugMatrix database
is_compound(compounds = c("aspirin", "paracetamol"))
}
}
