% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_one}
\alias{mat_one}
\title{Function for matrix of one}
\usage{
mat_one(n, m)
}
\arguments{
\item{n}{Number of rows}

\item{m}{Number of columns}
}
\value{
A matrix of ones5
}
\description{
The function \code{mat_one()} is used to generate a matrix of one.
}
\examples{
mat_one(3, 4)
}
