% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset&average-data.R
\name{mean_expression}
\alias{mean_expression}
\title{Average gene expression in different levels}
\usage{
mean_expression(
  ...,
  ge_matrix,
  metadata,
  probes = NULL,
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{probes}{A vector of Affymetrix probe IDs to be included. The default is \code{NULL},
which includes all probes..}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list of average data
}
\description{
The function \code{mean_expression()} is used to get average gene expression value in \code{group},
\code{compound}, \code{dose} and \code{time} levels.
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
avg_expr <- mean_expression(gr, ge_matrix = sim_data$expression, metadata = sim_data$metadata)
}
