% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy-probes-to-genes.R
\name{probes2genes}
\alias{probes2genes}
\title{Find gene information from probe set}
\usage{
probes2genes(affy_ids, organism = "rat")
}
\arguments{
\item{affy_ids}{Affymetric probe IDs}

\item{organism}{Sample organism used for experiment ("human" or "rat")}
}
\value{
A table of Ensembl gene symbol and other information
}
\description{
The function \code{probes2genes()} is used to get gene name, symbol, entrezid id etc from probes id.
}
\examples{
\dontrun{
affy_keys <- AnnotationDbi::keys(rat2302.db::rat2302.db, keytype = "PROBEID")
sample_id <- sample(affy_keys, 10)
probes2genes(affy_ids = sample_id, organism = "rat")
}
}
