% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{process_celfile}
\alias{process_celfile}
\title{Processing Gene Expression and Metadata of a Compound in the TG-GATEs Database}
\usage{
process_celfile(
  compound_path,
  fc = TRUE,
  method = "rma",
  multicore = FALSE,
  output_dir = missing_arg(),
  ...
)
}
\arguments{
\item{compound_path}{Full path to the compound directory, which should include
a folder named \code{celfile} containing CEL files, as well as a \code{tsv} attribute file.}

\item{fc}{A boolean indicating the format of the output gene expression data.
\itemize{
\item \code{TRUE} (default): The output will be in fold-change (FC) format.
\item \code{FALSE}: The output will be raw gene expression data.
}}

\item{method}{The normalization method to use, either \code{"rma"} or \code{"mas5"}. Default is \code{"rma"}.}

\item{multicore}{Used for parallel computing.
\itemize{
\item \code{FALSE} (default): No parallelization will be used.
\item \code{TRUE}: Computation will employ parallel processing, leaving one core for system use.
\item Integer: Specifies the number of workers for parallel processing.
}}

\item{output_dir}{Output directory for downloading data. If not specified,
the temporary path of the current R session will be used.}

\item{...}{Additional options for processing CEL files, passed on to
\link[affy:rma]{rma} or \code{\link[affy:mas5]{mas5}}, such as \code{normalize}, \code{background}, etc.}
}
\value{
A list containing:
\itemize{
\item \code{gexpr_data}: A matrix of gene expression values, with probes in rows and samples (CEL IDs) in columns.
\item \code{meta_data}: A data frame containing metadata about the samples and other related information.
}
}
\description{
The \code{process_celfile()} function generates gene expression data and associated metadata for samples.
This data is derived from CEL files and attribute files, respectively.
}
\examples{
compound <- "2NF"
\dontrun{
file_path <- tggates_rawdata(compound = "2-nitrofluorene")
proc_expr <- process_celfile(file_path)
head(proc_expr$gexpr_data)
head(proc_expr$meta_data)
}
}
\seealso{
Other Open TG-GATEs data download helpers: 
\code{\link{process_drugmatrix}()},
\code{\link{process_tggates}()},
\code{\link{tggates_rawdata}()}
}
\concept{Open TG-GATEs data download helpers}
