% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{process_tggates}
\alias{process_tggates}
\title{Process Compound Data from Open TG-GATEs}
\usage{
process_tggates(
  files,
  fc = TRUE,
  method = "rma",
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  ...
)
}
\arguments{
\item{files}{A vector of paths to compound directories, each containing a \code{celfile} folder
and a \code{.tsv} attribute file.}

\item{fc}{A boolean indicating the format of the output gene expression data.
\itemize{
\item \code{TRUE} (default): The output will be in fold-change (FC) format.
\item \code{FALSE}: The output will be raw gene expression data.
}}

\item{method}{The normalization method to use, either \code{"rma"} or \code{"mas5"}. Default is \code{"rma"}.}

\item{multicore}{Used for parallel computing.
\itemize{
\item \code{FALSE} (default): No parallelization will be used.
\item \code{TRUE}: Computation will employ parallel processing, leaving one core for system use.
\item Integer: Specifies the number of workers for parallel processing.
}}

\item{store}{Logical. If \code{TRUE} (the default), the resulting data will be saved as
two CSV files (gene expression and metadata) in the output directory.}

\item{output_dir}{Output directory for downloading data. If not specified,
the temporary path of the current R session will be used.}

\item{...}{Additional arguments passed to \code{process_celfile}.}
}
\value{
A list containing:
\itemize{
\item \code{expression}: A matrix of gene expression values, with probes in rows and samples (CEL IDs) in columns.
\item \code{metadata}: A data frame with attributes about samples and clinical information of compounds.
}
}
\description{
The \code{process_tggates()} function processes a set of compounds from the Open TG-GATEs
database. It handles all CEL and attribute files for each compound and generates two
CSV files: \code{expression.csv} and \code{metadata.csv}. If the \code{store} flag is set to \code{TRUE},
these CSV files will be saved in the specified output directory.
}
\examples{
\dontrun{
compound_list <- list("2-nitrofluorene", "N-nitrosomorpholine")
files <- lapply(compound_list, tggates_rawdata, species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "Single")
tgx_data <- process_tggates(files, store = TRUE)
head(tgx_data$expression)
head(tgx_data$metadata)
}
}
\seealso{
Other Open TG-GATEs data download helpers: 
\code{\link{process_celfile}()},
\code{\link{process_drugmatrix}()},
\code{\link{tggates_rawdata}()}
}
\concept{Open TG-GATEs data download helpers}
