% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate Random Samples from a Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Integer. The number of samples to generate.}

\item{alpha}{Numeric vector. The concentration parameters for the Dirichlet distribution.
Each element of \code{alpha} represents a parameter for the corresponding component of the Dirichlet distribution.}
}
\value{
A numeric vector of length \code{n * length(alpha)} representing random samples from the Dirichlet distribution.
}
\description{
This function generates random samples from a Dirichlet distribution using gamma function.
}
\examples{
# Generate a random sample from a Dirichlet distribution with three components
rdirichlet(1, c(100, 100, 100))

}
