% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi-network.R
\name{setup_stringdb}
\alias{setup_stringdb}
\title{Setup STRING Database}
\usage{
setup_stringdb(
  organism = c("human", "rat"),
  score_threshold = 200,
  version = "12"
)
}
\arguments{
\item{organism}{A character string specifying the sample organism used for the experiment. Options are "human" or "rat". Default is "rat".}

\item{score_threshold}{A numeric value specifying the threshold for the combined scores of the interactions. Default is 200.}

\item{version}{A character string specifying the version of the STRING database. Default is "11.5".}
}
\value{
An object of class \code{STRINGdb} representing the STRING database for the specified organism.
}
\description{
The \code{setup_stringdb()} function initializes the STRING database for a specified organism and score threshold.
}
\examples{
\dontrun{
# Setup STRING database for rat with a score threshold of 200
string_db <- setup_stringdb(organism = "rat", score_threshold = 200)
}
}
