% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{shrink_data}
\alias{shrink_data}
\title{Shorten data by removing data for compounds}
\usage{
shrink_data(
  compounds,
  ge_matrix,
  metadata,
  store = FALSE,
  output_dir = missing_arg()
)
}
\arguments{
\item{compounds}{A character vector of compound names to retrieve from the TG-GATES database.}

\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list.
\itemize{
\item A matrix of shortened gene expression values, with probes in rows and samples
(CEL IDs) in columns.
\item A data frame of shortened metadata containing sample information for compounds.
}
}
\description{
The \code{shirink_data()} function is used to remove data for one or more compounds from the
existing data.

The function \code{shrink_data()} is used to remove data for one or more compounds from the
existing data. It deletes corresponding samples of compounds from both the expression
and metadata tables, based on the database used.
}
\examples{
compounds <-  c("2NF", "NMOR", "MNU")
\dontrun{
otg_data <- get_tggates(compounds = c("2-nitrofluorene", "N-nitrosomorpholine", "N-methyl-N-nitrosourea"))
otg_srnkdata <- expand_data(compounds = c("2-nitrofluorene", "N-nitrosomorpholine"),
                           expr_data = otg_data$expression,
                           attr_data = otg_data$metadata,
                           database = "tggates")

head(otg_srnkdata$expression)
head(otg_srnkdata$metadata)
}
}
