% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-TGx-data.R
\name{sim_variance}
\alias{sim_variance}
\title{Simulate Variance Components for ICC}
\usage{
sim_variance(icc, n_component, equal_var = TRUE)
}
\arguments{
\item{icc}{Numeric value representing the intracluster correlation coefficient (ICC).}

\item{n_component}{Integer value indicating the number of variance components to simulate. Default is 3.}

\item{equal_var}{Logical value indicating whether the variance components should be equal (\code{TRUE}) or unequal (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
A numeric vector of length \code{n_component} representing the simulated variance components, normalized by the ICC.
}
\description{
This function generates simulated variance components based on a specified intracluster
correlation coefficient (ICC). It can simulate either equal or unequal variance components
across a specified number of components.
}
\examples{
# Simulate 3 variance components with equal variance
sim_variance(icc = 0.5, n_component = 3, equal_var = TRUE)

# Simulate 3 variance components with unequal variance
sim_variance(icc = 0.5, n_component = 3, equal_var = FALSE)

}
