% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-TGx-data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Gene Expression Data}
\usage{
simulate_data(
  n_gene,
  n_com,
  n_dose = 3,
  n_time = 4,
  n_rep = 3,
  mu = runif(1, -3, 3),
  d = 0.5,
  icc = 0.5
)
}
\arguments{
\item{n_gene}{Integer. Number of genes to simulate.}

\item{n_com}{Integer vector. Number of compounds in each group.}

\item{n_dose}{Integer. Number of dose levels. Default is 3.}

\item{n_time}{Integer. Number of time points. Default is 4.}

\item{n_rep}{Integer. Number of replicates for each condition. Default is 3.}

\item{mu}{Numeric. Mean expression level for the simulated genes. Default is a random value between -3 and 3.}

\item{d}{Numeric. Effect size for differential expression. Default is 0.5.}

\item{icc}{Numeric. Intracluster correlation coefficient. Default is 0.5.}
}
\value{
A list with two elements:
\describe{
\item{expression}{A matrix of simulated gene expression data. Rows represent genes and columns represent samples.}
\item{metadata}{A data frame containing metadata for each sample, including compound, dose level, time point, and replicate information.}
}
}
\description{
This function generates simulated gene expression data across multiple conditions,
including different compounds, dose levels, and time points.
}
\examples{
# Simulate gene expression data for 10 genes for 5 compounds in two groups.
sim_data <- simulate_data(n_gene = 10, n_com = c(5, 5))
head(sim_data$expression)
head(sim_data$metadata)

}
