% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-TGx-data.R
\name{simulate_ge}
\alias{simulate_ge}
\title{Simulate Gene Expression Data for Toxicogenomics Studies}
\usage{
simulate_ge(
  n_com,
  n_dose = 3,
  n_time = 4,
  n_rep = 3,
  mu = runif(1, -3, 3),
  d = 0.5,
  icc = 0.5,
  error_call = caller_env()
)
}
\arguments{
\item{n_com}{Numeric vector of length 2 indicating the number of compounds in each group
(e.g., toxic and non-toxic). Default is \code{c(5, 5)}.}

\item{n_dose}{Integer specifying the number of dose levels. Default is \code{3}.}

\item{n_time}{Integer specifying the number of time points. Default is \code{4}.}

\item{n_rep}{Integer specifying the number of replicates for each condition. Default is \code{3}.}

\item{mu}{Numeric value representing the overall mean of the gene expression. Default is a random value between -3 and 3.}

\item{d}{Numeric value for the effect size (Cohen's d) between the two groups. Default is \code{0.5}.}

\item{icc}{Numeric value for the intracluster correlation coefficient. It should be between 0 and 1. Default is \code{0.5}.}

\item{error_call}{Environment used to capture and handle errors with informative messages. Default is \code{caller_env()}.}
}
\value{
A numeric vector of simulated gene expression values.
}
\description{
This function generates simulated gene expression data based on specified parameters,
such as the number of compounds, doses, time points, and replicates. It also incorporates
a given effect size, intracluster correlation coefficient (ICC), and error variance.
}
\details{
The function simulates gene expression data by first generating variance components based on the provided ICC.
It then generates compound, dose, and time point effects using normal distributions.
}
\examples{
# Simulate gene expression data with default settings
simulated_data <- simulate_ge(n_com = c(3, 3))

# Simulate gene expression data with specific parameters
simulated_data <- simulate_ge(n_com = c(3, 3))

}
