% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-TGx-data.R
\name{simulate_tgxdata}
\alias{simulate_tgxdata}
\title{Simulate Toxicogenomics Expression Data}
\usage{
simulate_tgxdata(
  n_de = 10,
  n_ee = 10,
  n_com = c(5, 5),
  n_dose = 3,
  n_time = 4,
  n_rep = 3,
  mu = runif(1, -3, 3),
  d = 0.5,
  icc = 0.5
)
}
\arguments{
\item{n_de}{Integer. The number of differentially expressed (DE) genes to simulate.}

\item{n_ee}{Integer. The number of equivalently expressed (EE) genes to simulate.}

\item{n_com}{Integer vector. Number of compounds in each group.}

\item{n_dose}{Integer. Number of dose levels. Default is 3.}

\item{n_time}{Integer. Number of time points. Default is 4.}

\item{n_rep}{Integer. Number of replicates for each condition. Default is 3.}

\item{mu}{Numeric. Mean expression level for the simulated genes. Default is a random value between -3 and 3.}

\item{d}{Numeric. Effect size for differential expression. Default is 0.5.}

\item{icc}{Numeric. Intracluster correlation coefficient. Default is 0.5.}
}
\value{
A list containing:
\item{expression}{A matrix of simulated expression data for both DE and EE genes.}
\item{metadata}{A data frame of metadata associated with the simulated experiment.}
\item{gene_cl}{A vector indicating the class of each gene (DE or EE).}
}
\description{
This function generates simulated toxicogenomics (TGx) expression data, including differentially expressed (DE) genes
and equivalently expressed (EE) genes between groups. The function allows for customization of the number of DE and EE genes,
as well as various experimental conditions such as the number of compounds, doses, time points, and replicates.
}
\examples{
# Simulate TGx expression data with 20DEGs and 10 EEGs.
tgx_data <- simulate_tgxdata(n_de = 20, n_ee = 10, n_com = c(3, 3))

}
