% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-check.R
\name{test_column}
\alias{test_column}
\title{Validate Column Presence in Data Frame or List}
\usage{
test_column(column, df)
}
\arguments{
\item{column}{A character vector representing the name(s) of the column(s) to be checked.}

\item{df}{A data frame, tibble, or list in which the presence of the specified column(s) will be tested.}
}
\value{
The function does not return a value but throws an error if the column does not exist in the provided data structure.
}
\description{
This function checks whether a specified column exists in a provided data frame, tibble, or list. It provides informative error messages if the input does not meet the expected criteria.
}
\details{
\itemize{
\item If \code{df} is not a data frame, tibble, or list, an error is thrown.
\item If the \code{column} argument is empty or the specified column(s) do not exist in \code{df}, the function provides a detailed error message indicating the issue and suggesting possible corrections.
}
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9)
test_column("a", df) # No error
\dontrun{
test_column("d", df) # Throws an error: "d" does not exist in `df`
}
}
