% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-check.R
\name{test_data}
\alias{test_data}
\title{Validate Expression and Attribute Data}
\usage{
test_data(expr_data, attr_data, error_call = rlang::caller_env())
}
\arguments{
\item{expr_data}{A matrix or array containing expression data. The matrix should have \code{probes} as rows and \code{barcode} as columns.}

\item{attr_data}{A data frame or tibble containing the attribute data, with required columns including \code{barcode}, \code{compound_name}, \code{dose_level}, and \code{time_level}.}
}
\value{
This function does not return a value. It is used for validation and will throw an error if the data is not valid.
}
\description{
This function checks the validity of the expression data and attribute data provided by the user.
It ensures that the expression data is a matrix and that the attribute data is a data frame.
Additionally, it verifies that the column names in the expression data match the barcodes specified in the attribute data.
}
\details{
\itemize{
\item If \code{expr_data} is not a matrix or array, an error is thrown indicating that the expression data must be a matrix.
\item If \code{attr_data} is not a data frame, an error is thrown indicating that the attribute data must be a data frame.
\item The function also checks that all column names in \code{expr_data} are present in the \code{barcode} column of \code{attr_data}. If any are missing, an error is thrown.
}
}
\examples{
tgx_data <- simulate_tgxdata(n_de = 20, n_ee = 10, n_com = c(3, 3))
test_data(tgx_data$expression, tgx_data$metadata)
\dontrun{
# Example with invalid data
expr_data_invalid <- array(rnorm(100), dim = c(10, 10, 1))
attr_data_invalid <- list(barcode = paste0("sample", 1:10))
test_data(expr_data_invalid, attr_data_invalid) # This will throw an error
}

}
