% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-check.R
\name{test_element}
\alias{test_element}
\title{Validate Input Against a List of Elements}
\usage{
test_element(input, elements, error_call = caller_env())
}
\arguments{
\item{input}{A vector representing the input to be validated against the \code{elements}.}

\item{elements}{A vector of valid elements that \code{input} should be compared against.}

\item{error_call}{The environment to be used in the \code{cli_abort} call for error reporting. Default is \code{caller_env()}.}
}
\value{
This function does not return a value. It is used for validation and will throw an error if validation fails.
}
\description{
This function checks whether a given \code{input} is present in a specified list of \code{elements}.
If the input is missing, empty, or contains elements not found in the list, the function
throws an informative error message using \code{cli_abort}.
}
\examples{
valid_elements <- c("apple", "banana", "cherry")
test_element("apple", valid_elements)  # No error
\dontrun{
test_element("orange", valid_elements) # Throws an error
}
}
