% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-check.R
\name{test_input}
\alias{test_input}
\title{Validate and Auto-Correct Function Inputs}
\usage{
test_input(input, inputs, auto_input = FALSE, error_call = caller_env())
}
\arguments{
\item{input}{The input provided by the user to be validated.}

\item{inputs}{A vector of valid inputs to compare against. If \code{inputs} is missing, the function will attempt to retrieve the default inputs from the parent environment.}

\item{auto_input}{Logical value indicating whether to automatically correct the input to the first valid entry in \code{inputs} if multiple or invalid inputs are provided. Default is \code{FALSE}.}

\item{error_call}{The environment in which to evaluate the error call. Default is \code{caller_env()}.}
}
\value{
This function does not return a value. It either validates the input or throws an error if the input is invalid.
}
\description{
This function validates the input provided by the user against a list of expected inputs.
It can automatically correct the input if \code{auto_input} is set to \code{TRUE},
or it will prompt the user with an error message if the input is invalid.
}
\examples{
# Assume valid inputs are "A", "B", and "C"
valid_inputs <- c("A", "B", "C")
test_input(input = "A", inputs = valid_inputs)

}
