% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{tggates_link}
\alias{tggates_link}
\title{Download Link for Data in the Open TG-GATEs Database}
\usage{
tggates_link(
  compound,
  species = c("Rat", "Human"),
  data_type = c("in_vivo", "in_vitro"),
  tissue = c("Liver", "Kidney"),
  dose_type = c("Single", "Repeat")
)
}
\arguments{
\item{compound}{An available compound name.}
}
\value{
A download link.
}
\description{
The \code{tggates_link()} function retrieves the download link for perturbation data of compounds along with metadata from the \href{https://toxico.nibiohn.go.jp/english/}{TG-GATEs} database.
}
\section{Open TG-GATEs Database}{

The \href{https://toxico.nibiohn.go.jp/english/}{TG-GATEs} database includes tests on 170
distinct compounds, primarily medicinal drugs, across both \emph{in-vivo} and \emph{in-vitro}
experimental setups. For \emph{in-vivo} trials, male Sprague Dawley rats were used, and
the studies were organized into single-dose and repeated-dose protocols. Additionally,
the database includes two \emph{in-vitro} studies utilizing primary hepatocytes from male
Sprague Dawley rats and human donors.
}

\examples{
# Download link for downloading perturbation data of Acetaminophen for different experimental settings
tggates_link(compound = "acetaminophen", species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "single")
tggates_link(compound = "acetaminophen", species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "Repeat")
tggates_link(compound = "acetaminophen", species = "Rat", data_type = "in_vitro", tissue = "Liver", dose_type = "Single")
tggates_link(compound = "acetaminophen", species = "Human", data_type = "in_vitro", tissue = "Liver", dose_type = "Single")
}
\seealso{
Other Open TG-GATEs Data Download Helpers: 
\code{\link{tggates_compounds}()}
}
\concept{Open TG-GATEs Data Download Helpers}
