% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicogenomics-data.R
\name{tggates_rawdata}
\alias{tggates_rawdata}
\title{Download Microarray and Metadata for Query Compounds from the TG-GATEs Database}
\usage{
tggates_rawdata(
  compound,
  species = c("Rat", "Human"),
  data_type = c("in_vivo", "in_vitro"),
  tissue = c("Liver", "Kidney"),
  dose_type = c("Single", "Repeat"),
  output_dir = missing_arg()
)
}
\arguments{
\item{compound}{An available compound name.}

\item{output_dir}{Output directory for downloading data. If not specified,
the temporary path of the current R session will be used.}
}
\value{
A folder containing \code{CEL} files and a \code{tsv} metadata file.
}
\description{
The \code{tggates_rawdata()} function downloads Affymetrix CEL files and associated attribute data,
including a data dictionary, clinical, and biochemical information, for a specific compound
from the \href{https://toxico.nibiohn.go.jp/english/}{TG-GATEs} database. The data will be
organized in a directory named \strong{compound_name.data_type} in the specified output location.
Within this directory, a sub-folder named \strong{celfiles} will store all CEL files, and a .tsv
file containing attribute information will also be created.

The function ensures that gene expression data, measured at various doses and time points,
are systematically stored. The download speed will depend on your internet connection.
}
\examples{
\dontrun{
tggates_rawdata(compound = "2-nitrofluorene", species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "Single")
tggates_rawdata(compound = "N-nitrosomorpholine", species = "Rat", data_type = "in_vivo", tissue = "Liver", dose_type = "Single")
}
}
\seealso{
Other Open TG-GATEs data download helpers: 
\code{\link{process_celfile}()},
\code{\link{process_drugmatrix}()},
\code{\link{process_tggates}()}
}
\concept{Open TG-GATEs data download helpers}
