% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm-degs.R
\name{tgx_degs}
\alias{tgx_degs}
\title{Identification of causal differentially expressed genes}
\usage{
tgx_degs(
  ...,
  ge_matrix,
  metadata,
  p_cutoff = 0.05,
  p_adjust = "none",
  gr_diff = FALSE,
  log10p = FALSE,
  multicore = FALSE,
  store = FALSE,
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{p_cutoff}{P-value cutoff threshold.}

\item{p_adjust}{Adjusted \code{p-values} using Bonferroni methods. default is \code{TRUE}.}

\item{gr_diff}{Show the column(s) for group mean differences. The default is \code{FALSE}.}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A data frame of gene identification result.
}
\description{
The function \code{tgx_genes()} is used to find final gene set using Leave One Out Compound (LOOC) method.
Genes misrepresented by a compound were removed using Leave One Out Compound (LOOC) method.
In LOOC process, each compound is removed from the experiment and find differentially expressed (DE)
genes for every LOO. The final gene set is therefore the intersection of the LOOC genes. Genes which
are co-regulated by a specific compound then filter out in LOOC process.
}
\examples{
sim_data <- simulate_tgxdata(n_de = 10, n_ee = 10, n_com = c(5,5))
gr <- list(A = paste0("Compound", 1:5), B = paste0("Compound", 6:10))
tgx_degs(gr, ge_matrix = sim_data$expression, metadata = sim_data$metadata)
}
