% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-data.R
\name{update_data}
\alias{update_data}
\title{Clean and refresh the data}
\usage{
update_data(
  compounds,
  ge_matrix,
  metadata,
  probes = NULL,
  multicore = FALSE,
  store = FALSE,
  method = "rma",
  output_dir = missing_arg(),
  error_call = caller_env()
)
}
\arguments{
\item{compounds}{A character vector of compound names to retrieve from the TG-GATES database.}

\item{ge_matrix}{A matrix of gene expression data for given compounds.}

\item{metadata}{Metadata for given compounds, usually in the form of a data frame or
tibble object.}

\item{probes}{A vector of Affymetrix probe IDs to be included. The default is \code{NULL},
which includes all probes..}

\item{multicore}{A logical value indicating whether to use multiple cores for processing. Default is \code{FALSE}.}

\item{store}{A logical value indicating whether to store the processed data. Default is \code{FALSE}.}

\item{method}{A character string specifying the preprocessing method to use. Default is \code{"rma"}.}

\item{output_dir}{A character string specifying the directory to store output files. Default is \code{missing_arg()}.}
}
\value{
A list.
\itemize{
\item A matrix of gene update expression values, probes in row and sample (CEL ID) in column.
\item A data frame of update metadata about sample information of given compounds.
}
}
\description{
The \code{update_data()} function updates the given gene expression and metadata by adding or
removing data for compounds based on queried compounds.

The function \code{update_data()} is used to add or remove data for compounds in the existing
dataset. The compounds you wish to add must be available in the specified database.
The available compounds in \href{https://toxico.nibiohn.go.jp/english/}{TG-GATEs} and
\href{https://ntp.niehs.nih.gov/data/drugmatrix}{DrugMatrix} databases can be found by
calling the functions \code{\link[=tggates_compounds]{tggates_compounds()}} and \code{\link[=dm_compounds]{dm_compounds()}}, respectively.
Furthermore, the compounds you wish to remove must have both expression and attribute data.
}
\examples{
\dontrun{
otg_data <- get_tggates(compounds = c("2-nitrofluorene", "N-nitrosomorpholine"))
otg_cnrdata <- update_data(compounds = c("2-nitrofluorene", "N-methyl-N-nitrosourea"),
                        ge_matrix = otg_data$expression,
                        metadata = otg_data$metadata)
head(otg_cnrdata$expression)
head(otg_cnrdata$metadata)
}
}
\seealso{
\code{\link[=expand_data]{expand_data()}} for expanding the data by adding data for query compounds,
\code{\link[=shrink_data]{shrink_data()}} for reducing the data by removing data for unwanted compounds.
}
