% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{download.file}
\alias{unzip}
\alias{packageVersion}
\alias{ReadAffy}
\alias{mas5}
\alias{rma}
\alias{exprs}
\alias{median}
\alias{across}
\alias{pull}
\alias{distinct}
\alias{ungroup}
\alias{arrange}
\alias{group_by}
\alias{select}
\alias{filter}
\alias{rename}
\alias{rename_with}
\alias{mutate_at}
\alias{as_tibble}
\alias{write_csv}
\alias{detectCores}
\alias{makeCluster}
\alias{clusterExport}
\alias{clusterCall}
\alias{stopCluster}
\alias{registerDoParallel}
\alias{getDoParName}
\alias{getDoParWorkers}
\alias{registerDoSEQ}
\alias{is_empty}
\alias{.data}
\alias{englue}
\alias{list2}
\alias{is_bare_list}
\alias{inherits_any}
\alias{is_missing}
\alias{missing_arg}
\alias{cli_alert_warning}
\alias{cli_abort}
\alias{style_bold}
\alias{cli_ul}
\alias{bg_magenta}
\alias{bg_red}
\alias{col_red}
\alias{col_green}
\alias{cli_alert_info}
\alias{col_br_red}
\alias{cli_alert_success}
\alias{backtick}
\alias{glue}
\alias{glue_collapse}
\alias{nest}
\alias{pivot_wider}
\alias{tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{affy}{\code{\link[affy]{mas5}}, \code{\link[affy:read.affybatch]{ReadAffy}}, \code{\link[affy]{rma}}}

  \item{Biobase}{\code{\link[Biobase]{exprs}}}

  \item{cli}{\code{\link[cli:ansi-styles]{bg_magenta}}, \code{\link[cli:ansi-styles]{bg_red}}, \code{\link[cli:ansi-styles]{bg_red}}, \code{\link[cli]{cli_abort}}, \code{\link[cli:cli_alert]{cli_alert_info}}, \code{\link[cli:cli_alert]{cli_alert_success}}, \code{\link[cli:cli_alert]{cli_alert_warning}}, \code{\link[cli]{cli_ul}}, \code{\link[cli:ansi-styles]{col_br_red}}, \code{\link[cli:ansi-styles]{col_green}}, \code{\link[cli:ansi-styles]{col_red}}, \code{\link[cli:ansi-styles]{style_bold}}}

  \item{doParallel}{\code{\link[doParallel]{registerDoParallel}}}

  \item{dplyr}{\code{\link[dplyr]{across}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{foreach}{\code{\link[foreach:getDoParWorkers]{getDoParName}}, \code{\link[foreach]{getDoParWorkers}}, \code{\link[foreach]{registerDoSEQ}}}

  \item{glue}{\code{\link[glue:quoting]{backtick}}, \code{\link[glue]{glue}}, \code{\link[glue]{glue_collapse}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{parallel}{\code{\link[parallel:clusterApply]{clusterCall}}, \code{\link[parallel:clusterApply]{clusterExport}}, \code{\link[parallel]{detectCores}}, \code{\link[parallel]{makeCluster}}, \code{\link[parallel:makeCluster]{stopCluster}}}

  \item{readr}{\code{\link[readr:write_delim]{write_csv}}}

  \item{rlang}{\code{\link[rlang:dot-data]{.data}}, \code{\link[rlang]{englue}}, \code{\link[rlang]{inherits_any}}, \code{\link[rlang:bare-type-predicates]{is_bare_list}}, \code{\link[rlang]{is_empty}}, \code{\link[rlang:missing_arg]{is_missing}}, \code{\link[rlang]{list2}}, \code{\link[rlang]{missing_arg}}}

  \item{stats}{\code{\link[stats]{median}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}

  \item{tidyr}{\code{\link[tidyr]{nest}}, \code{\link[tidyr]{pivot_wider}}}

  \item{utils}{\code{\link[utils]{download.file}}, \code{\link[utils:packageDescription]{packageVersion}}, \code{\link[utils]{unzip}}}
}}

