% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-processsing.R
\name{start_parallel}
\alias{start_parallel}
\title{Starting parallel computing}
\usage{
start_parallel(multicore = TRUE, error_call = caller_env(), ...)
}
\arguments{
\item{multicore}{Parallel computation (default is \code{TRUE})}

\item{...}{Additional parameter}
}
\value{
Class of parallel object
}
\description{
The function \code{start_parallel()} is used to setup parallel processing by appointing multiple
workers for a sequence of similar task.
}
\examples{
\dontrun{
start_parallel(multicore = TRUE)
}
}
