% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-processsing.R
\name{stop_parallel}
\alias{stop_parallel}
\title{Stop parallel setup}
\usage{
stop_parallel(cluster, ...)
}
\arguments{
\item{cluster}{Class of parallel object}

\item{...}{Additional parameter}
}
\value{
Exit cluster
}
\description{
The function \code{stop_parallel()} is used to stop parallel processing.
}
\examples{
\dontrun{
cll <- start_parallel()
stop_parallel(attr(cll, "cluster"))
}
}
